Refactored and moved old navbar file into components/Navbar.jsx to be in line with other changes
