+++ src/components/Navbar.css
@@ -1,6 +1,3 @@
-/* Navbar.css */
-
-/* Navbar container */
 .navbar {
     background-color: #27AE60; /* Green background */
     padding: 10px 20px; /* Spacing inside the navbar */
